fix: validation functions on separate file from contact script
