+++ index.html
@@ -258,5 +258,6 @@
     <script type="module" src="./scripts/newsletter.js"></script>
     <script src="./scripts/currencies.js"></script>
     <script src="./scripts/slider.js"></script>
+    <script type="module" src="./scripts/contact.js"></script>
 </body>
 </html>
