+++ scripts/formValidation.js
@@ -34,31 +34,14 @@ export const submitForm = (userName, userEmail) => {
             'Content-type': 'application/json; charset=UTF-8',
         },
         })
-        .then((response) => response.json())
+        .then((response) => {
+            if (response.ok) {
+                response.json().catch((error) => console.error('Error:', error));
+            }
+        })
         .then((json) => { 
             console.log(json); 
             alert("Form submitted successfully")
         }) 
         .catch((error) => console.error('Error:', error));
 }
-
-document.getElementById("contact").addEventListener("submit", (event) => {
-    const nameInput = document.getElementById("contact-name");
-    const userName = nameInput.value;
-    const emailInput = document.getElementById("contact-email");
-    const userEmail = emailInput.value;
-    const checkbox = document.getElementById("contact-policies-check");
-
-    event.preventDefault();
-
-    const checked = checkValues(nameInput, userName, emailInput, userEmail, checkbox);
-    if (checked) {
-        submitForm(userName, userEmail);
-        nameInput.classList.remove("input--rejected");
-        emailInput.classList.remove("input--rejected");
-        checkbox.classList.remove("policies-check--rejected");
-        nameInput.value = "";
-        emailInput.value = "";
-        checkbox.checked = false;
-    }
-})
